requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
boto3==1.18.12
botocore==1.21.12
bs4==0.0.1
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
cryptography==3.4.7
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.1.7
django-allauth==0.45.0
django-ckeditor==6.1.0
django-cors-headers==3.7.0
django-filter==2.4.0
django-js-asset==1.2.2
django-storages==1.11.1
django-taggit==1.5.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.7.1
fontawesome-free==5.15.4
gunicorn==20.1.0
idna==3.2
jmespath==0.10.0
Markdown==3.3.4
oauthlib==3.1.1
Pillow==8.1.2
psycopg2-binary==2.9.1
pycparser==2.20
PyJWT==2.1.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
s3transfer==0.5.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
urllib3==1.26.6
whitenoise==5.3.0
